Add UUID to user storage and enable user search by UUID
Improved the user storage model by adding UUID. The uuid is now included while setting the user and changes were made in order to allow the retrieval of users by UUID. This provides an additional method of identifying and getting user details. The Azure auth service has also been added to the ServicesIntegrationProcedures.