Add DiscordRpc integration to the ServicesIntegrationProcedures
This commit introduces a new feature, the integration of a Discord RPC client to the ServicesIntegrationProcedures. It involves adding methods for setting and getting a Discord RPC client, a constant for storing a reference to the client, and designing a new interface describing the properties of a Discord RPC client.