Update profile functionality with icon and background image support
Updated the UpdateProfile method in IProfileProcedures to support the addition of both an icon and background image. Introduced a new method for file conversion to base64 format. Also added a new property to the IGameProfile interface and its implementation to store the key of the new background image. Furthermore, IFileStorageProcedures is now equipped to handle file upload.