Refactored FileStorageProcedures in GmlCore
The codebase up-gradation process was involved in this commit. Unnecessary import statements were removed to clean up the code. Further, added an else clause to handle the case when the metadata doesn't exist and revised the logic for putting objects, including changes to bucket name and content type, as well as added object size. This provided better handling when dealing with different situations.