Add function to get all versions
A new function 'GetAllVersions' has been added to the 'GameDownloaderProcedures' class which fetches all versions of the game and introduced the IVersion interface and MineVersion implementation. This function will return a list of all the available versions of the game, each of them represented by the 'MineVersion' class. Also updated related test cases.