Update property names and enhance version control in core interfaces
This commit updates the property 'TextureUrl' to 'TextureSkinUrl' in IUser interface and its implementation. Also, it enhances the version control in core interfaces by adding a new method 'GetAllowVersions' in 'IGameDownloaderProcedures'. This method will return allowed versions depending on the game loader type and Minecraft version. The corresponding changes are also applied in the 'GameDownloaderProcedures' implementation.