Refactor and enhance profile loading mechanism
ProfileLoaderStateMachine was removed and the responsibility for profiling state was added to the game profile itself. This ensures a more clean and intuitive control of profile states. Multiple methods were also updated to streamline the downloading and restoration of game profiles. A method for creating a mods folder was implemented. Exception handing was improved throughout.