Update to v0.1.0-beta4 (#53)
* Remove redundant Clear method from NotificationProcedures

The Clear method was redundant and not being utilized anywhere in the codebase. Its removal streamlines the class and ensures consistency in the notification handling procedures.

* Refactor whitelist methods in ProfileProcedures

Converted async methods for adding and removing files and folders to non-blocking tasks. Also added bulk operations for adding and removing folders in IProfileProcedures interface to improve performance and maintainability.

* Update submodule link CmlLib.Core.Installer.Forge

* Update Minecraft server addresses in unit tests

Replaced old server addresses with new ones in multiple unit tests to reflect the current server IPs. Commented out a redundant assertion in the GmlManager test and added a TODO note to fix the endpoint.

* Refactor unit tests to use dynamic profile names

Replaced hardcoded profile names with a dynamic variable `name` to make the tests more flexible and maintainable. Additionally, commented out a redundant assertion in the installation test method to clean up the code.

* #52 Use case-insensitive comparison for profile names

Replaced the exact match check with a case-insensitive comparison when adding new profiles. This prevents duplicate profiles with names that differ only by case.

* Handle DirectoryNotFoundException in GameDownloader

Add specific catch block for DirectoryNotFoundException to provide a detailed warning message and log the unsupported operating system profile creation attempt. This ensures clearer feedback and better error handling for unsupported systems.

* Add support for game-specific arguments in process creation

Enhanced the profile and game downloader procedures to include game-specific arguments. Updated model interfaces and implementations to support these new arguments and ensured backward compatibility with existing JVM arguments.

* Add support for game arguments in profile procedures

Extended profile management functionality to include game arguments. Updated method signatures and relevant handlers to accept and process the new game arguments parameter. This ensures profiles can now be configured with specific game arguments alongside existing JVM arguments.

* Add GameArguments property to GameProfileInfo

This new property is meant to store game-specific arguments separately from JVM arguments. It increases the clarity and flexibility of the GameProfileInfo class.

* Add GameArguments property to GameProfileInfo

This new property is meant to store game-specific arguments separately from JVM arguments. It increases the clarity and flexibility of the GameProfileInfo class.

* Ensure file existence in LocalStorage verification

Previously, the code only checked for a non-null localFileInfo. The update adds an additional check to confirm that the file actually exists on the filesystem before proceeding. This helps prevent potential runtime errors and ensures that file operations are only attempted on valid files.

* Optimize game arguments handling in profile procedures

Updated the game arguments addition to use AddRange with the split method. This ensures that individual words in the arguments string are correctly separated and added, improving the handling of game arguments in the profile procedures.

* Add cache restoration flag to GetAllProfileFiles method

Updated the GetAllProfileFiles method across multiple classes to include an optional needRestoreCache parameter. This change enables conditional cache restoration and improves cache handling during file retrieval processes.

* Update submodule link CmlLib.Core.Installer.Forge

---------

Co-authored-by: Gru <78332542+gru2007@users.noreply.github.com>
Co-authored-by: akemiko <tcop29711@gmail.com>