Feature/texture protocol (#65)
* Add TextureProtocol support to storage settings

Updated the storage settings to include support for TextureProtocol. Modified related methods and interfaces to accommodate the new TextureProtocol property.

* Remove default TextureProtocol initialization

Removed the default initialization value for the `TextureProtocol` in `StorageSettings` and explicitly set values for the `TextureProtocol` enum. This ensures the protocol must be explicitly defined, improving code clarity and avoiding potential confusion with default values.

---------

Co-authored-by: Akemiko <tcop29711@gmail.com>